+++ .github/workflows/multi-os-build.yml
@@ -41,7 +41,7 @@ jobs:
         uses: gradle/gradle-build-action@v2.2.2
         with:
           cache-read-only: ${{ github.ref != 'refs/heads/main' }}
-          arguments: buildPlugin --no-daemon
+          arguments: build buildPlugin --no-daemon
 
 
       # Collect Tests Result of failed tests
