+++ .github/workflows/publish.yml
@@ -11,6 +11,7 @@ jobs:
 
   # Prepare and publish the plugin to the Marketplace repository
   publish:
+#    todo: if: "github.event.release.released"
     name: Publish Plugin
     runs-on: ubuntu-latest
     steps:
