+++ build.gradle.kts
@@ -93,7 +93,10 @@ tasks {
     wrapper {
         gradleVersion = properties("gradleVersion")
         distributionType = Wrapper.DistributionType.ALL
-        distributionBase = Wrapper.PathBase.PROJECT
+        distributionBase = Wrapper.PathBase.GRADLE_USER_HOME
+        distributionPath = "wrapper/dists"
+        archiveBase = Wrapper.PathBase.GRADLE_USER_HOME
+        archivePath = "wrapper/dists"
     }
 
     patchPluginXml {
@@ -153,12 +156,10 @@ tasks {
         //rider contributes to prepareSandbox, so it needs to run before runIde
         dependsOn("prepareSandbox")
         dependsOn(":rider:prepareSandboxForRider")
-        maxHeapSize = "4g"
+        maxHeapSize = "2g"
         // Rider's backend doesn't support dynamic plugins. It might be possible to work with auto-reload of the frontend
         // part of a plugin, but there are dangers about keeping plugins in sync
         autoReloadPlugins.set(false)
-        //todo: workaround, it's a Gradle IntelliJ Plugin issue, it will be fixed in version 1.8.0 according to jetbrains developers.
-        jvmArgs = listOf("--add-opens=java.desktop/sun.awt.X11=ALL-UNNAMED")
     }
