+++ buildSrc/build.gradle.kts
@@ -12,7 +12,7 @@ repositories {
 
 dependencies {
     //Note: gradle-intellij-plugin 1.5.3 has some bugs with runIde that does not detect changes in kotlin UI DSL
-    implementation("org.jetbrains.intellij.plugins:gradle-intellij-plugin:1.7.0")
+    implementation("org.jetbrains.intellij.plugins:gradle-intellij-plugin:1.8.0")
     implementation("org.jetbrains.kotlin:kotlin-gradle-plugin:1.7.10")
     implementation("com.dorongold.plugins:task-tree:2.1.0")
     implementation("com.glovoapp.semantic-versioning:com.glovoapp.semantic-versioning.gradle.plugin:1.1.10")
@@ -29,13 +29,6 @@ kotlin {
     }
 }
 
-/*
-todo: upgrade gradle to 7.5
-this warning can be ignored:
-'compileJava' task (current target is 11) and 'compileKotlin' task (current target is 1.8) jvm target compatibility should be set to the same Java version.
-gradle issue:
-https://github.com/gradle/gradle/issues/18935
- */
 
 tasks {
