+++ buildSrc/src/main/kotlin/common-java-library.gradle.kts
@@ -1,6 +1,10 @@
 import org.gradle.api.tasks.testing.logging.TestExceptionFormat
 import org.gradle.api.tasks.testing.logging.TestLogEvent
 
+//this plugin is for projects that are pure java, meaning they don't have the
+//org.jetbrains.intellij plugin.
+//for example analytics-provider, model
+
 plugins {
     id("digma-base")
     `java-library`
@@ -8,69 +12,7 @@ plugins {
 
 
 tasks {
-    create("buildPlugin") {
-        //a workaround: if the project is built by calling buildPlugin
-        // then build is not called for projects that are not intellij plugin project.
-        dependsOn(build)
-    }
-
-
-    withType<Test> {
-        doFirst {
-            logger.lifecycle("${project.name}:${name}: testing java with {}", javaLauncher.get().executablePath)
-        }
-
-        addTestListener(object : TestListener {
-            override fun beforeTest(testDescriptor: TestDescriptor) {}
-            override fun afterTest(testDescriptor: TestDescriptor, result: TestResult) {}
-
-            override fun beforeSuite(suite: TestDescriptor) {
-                if (suite.parent == null) { // root suite
-                    logger.lifecycle("Starting Test suite {}", suite)
-                }
-            }
-
-            override fun afterSuite(suite: TestDescriptor, result: TestResult) {
-                if (suite.parent == null) { // root suite
-                    logger.lifecycle(
-                        "Test suite ${suite.name} completed:: ${result.resultType}, " +
-                                "success ${result.successfulTestCount}, " +
-                                "failed ${result.failedTestCount}, " +
-                                "skipped ${result.skippedTestCount}."
-                    )
-
-                }
-            }
-        })
-
-
-        testLogging {
-            lifecycle {
-                events = mutableSetOf(TestLogEvent.FAILED)
-                exceptionFormat = TestExceptionFormat.SHORT
-                showExceptions = true
-                showCauses = true
-                showStackTraces = false
-                showStandardStreams = false
-            }
-            debug {
-                events = mutableSetOf(TestLogEvent.FAILED, TestLogEvent.PASSED, TestLogEvent.SKIPPED)
-                exceptionFormat = TestExceptionFormat.FULL
-                showExceptions = true
-                showCauses = true
-                showStackTraces = true
-                showStandardStreams = true
-            }
-            info.events = debug.events
-            info.exceptionFormat = debug.exceptionFormat
-
-
-        }
-
-        addTestOutputListener { testDescriptor, outputEvent ->
-            if (outputEvent.destination == TestOutputEvent.Destination.StdErr) {
-                logger.error("Test: " + testDescriptor + ", error: " + outputEvent.message)
-            }
-        }
-    }
+    //a workaround: if the project is built by calling buildPlugin
+    // then build is not called for projects that are not intellij plugin project.
+    create("buildPlugin").dependsOn(build)
 }
