+++ buildSrc/src/main/kotlin/common-kotlin.gradle.kts
@@ -23,9 +23,8 @@ plugins {
 
 kotlin {
     jvmToolchain {
-            (this as JavaToolchainSpec).languageVersion.set(JavaLanguageVersion.of(properties("javaVersion", project)))
-        //to build with amazon corretto download it in the gradlew script. see resharper-unity
-        //vendor.set(JvmVendorSpec.AMAZON)
+            this.languageVersion.set(JavaLanguageVersion.of(properties("javaVersion", project)))
+            this.vendor.set(JvmVendorSpec.AMAZON)
     }
 }
