+++ gradle.properties
@@ -27,7 +27,7 @@ platformPlugins =
 javaVersion = 11
 
 # Gradle Releases -> https://github.com/gradle/gradle/releases
-gradleVersion = 7.4.2
+gradleVersion=7.5.1
 
 # Opt-out flag for bundling Kotlin standard library.
 # See https://plugins.jetbrains.com/docs/intellij/kotlin.html#kotlin-standard-library for details.
