+++ model/build.gradle.kts
@@ -18,7 +18,3 @@ dependencies {
 
     testImplementation(kotlin("test"))
 }
-
-tasks.test {
-    useJUnitPlatform()
-}
