+++ src/main/kotlin/org/digma/intellij/plugin/ui/list/insights/HotspotPanel.kt
@@ -6,6 +6,7 @@ import org.digma.intellij.plugin.ui.list.PanelsLayoutHelper
 import javax.swing.JPanel
 
 
+@Suppress("UNUSED_PARAMETER")
 fun hotspotPanel(modelObject: HotspotInsight, panelsLayoutHelper: PanelsLayoutHelper): JPanel {
     return createInsightPanel(
         "This is an error hotspot", "Many major errors occur or propagate through this function",
